+++ NEWS.md
@@ -3,6 +3,7 @@
 ## Changes in 0.0.3 [2017-01-06]
 
 - Fix `metOxF == 1` for peptides without methionine.
+- Add proline rule (Parker 2002, rule 14).
 
 ## Changes in 0.0.2 [2017-01-04]